+++ src/parser/rogue/outlaw/CombatLogParser.js
@@ -16,7 +16,7 @@ import Energy from './modules/core/Energy';
 import SpellEnergyCost from '../shared/resources/SpellEnergyCost';
 
 import RestlessBlades from './modules/core/RestlessBlades';
-import SliceAndDiceUptime from './modules/spells/SliceAndDiceUptime';
+import SliceAndDiceUptime from './modules/talents/SliceAndDiceUptime';
 
 class CombatLogParser extends CoreCombatLogParser {
   static specModules = {
@@ -37,13 +37,13 @@ class CombatLogParser extends CoreCombatLogParser {
     spellEnergyCost: SpellEnergyCost,
 
     //Core
-    sliceAndDiceUptime: SliceAndDiceUptime,
     restlessBlades: RestlessBlades,
     //Items
 
     //Casts
-    
+
     //Talents
+    sliceAndDiceUptime: SliceAndDiceUptime,
 
     // Outlaw's throughput benefit isn't as big as for other classes since we don't have a lot of free gcds to use
     arcaneTorrent: [ArcaneTorrent, {
